Create Script_for_LogFile.py
IT AUTOMATION WITH PYTHON:-
Task 1:-
Daily report that tracks the use of machines to know which users are currently connected to which machines, it collects every event that happens on the machines on the network. Among the many events collected it records each time a user logs in or out of a computer.
This script generates a report of which users are logged in to which machines at that time.